API: Add 6h records/gains
